refactor: 重构ToolArgsValue，使用泛型约束简化类型安全访问
重构内容：

1. ToolArgsValue函数重构：
   ```go
   // 重构前：复杂的类型转换逻辑
   func ToolArgsValue[T any](args ToolArgs, key string, defaultValue T) T {
       value, exists := args[key]
       if !exists {
           return defaultValue
       }
       return convertValue(value, defaultValue) // 复杂的类型转换
   }

   // 重构后：使用泛型约束，简洁的类型安全访问
   type Primitive interface {
       ~string | ~float64 | ~int64 | ~bool | ~[]string
   }

   func ToolArgsValue[T Primitive](args ToolArgs, key string, defaultValue T) T {
       if value, ok := args[key]; ok {
           if typedValue, ok := value.(T); ok {
               return typedValue
           }
       }
       return defaultValue
   }
   ```

2. 移除复杂的convertValue函数：
   - 移除strconv依赖
   - 移除复杂的类型转换逻辑
   - 简化代码结构

3. 代码格式化改进：
   ```go
   // 对齐注释
   maxAttempts := 1 // max attempts 1 means no retry

   // 清晰的切片表达式
   result = result[0 : len(result)-1]

   // 移除多余空行
   ```

重构优势：

1. 类型安全：
   - 使用Go泛型约束Primitive interface
   - 只支持5种基本类型：string、float64、int64、bool、[]string
   - 编译时类型检查，避免运行时类型错误

2. 代码简化：
   - 从复杂的类型转换逻辑 → 简洁的类型断言
   - 移除convertValue函数和strconv依赖
   - 代码行数大幅减少

3. 性能提升：
   - 直接类型断言，无需复杂的类型判断
   - 减少函数调用层级
   - 更高效的代码执行

4. 可维护性：
   - 清晰的类型约束定义
   - 简单的实现逻辑
   - 易于理解和修改

类型约束设计：

```go
// Primitive types tool arguments support:
// 1. string   - string
// 2. float64  - number
// 3. int64    - integer
// 4. bool     - boolean
// 5. []string - array
type Primitive interface {
    ~string | ~float64 | ~int64 | ~bool | ~[]string
}
```

设计决策：

1. 限制支持类型：
   - 优点：类型安全，避免错误
   - 缺点：不支持其他类型（如int、[]int等）
   - 权衡：对于工具参数，这5种类型基本够用

2. 使用类型断言：
   - 直接检查类型是否匹配
   - 不进行类型转换，保持数据原样
   - 避免转换过程中的精度损失或错误

3. 编译时检查：
   - 泛型约束在编译时检查类型
   - 避免运行时类型错误
   - 提高代码可靠性

测试重构：

1. 测试用例重构：
   ```go
   // 5种类型的全面测试：
   // 1. string类型测试
   // 2. float64类型测试
   // 3. int64类型测试
   // 4. bool类型测试
   // 5. []string数组类型测试
   ```

2. 测试覆盖：
   - 每种类型的基本情况
   - 默认值处理
   - 类型不匹配的情况

3. 测试清晰度：
   - 独立的测试用例
   - 清晰的测试名称
   - 易于理解的断言

其他改进：

1. 代码格式化：
   - 对齐注释和代码
   - 改善代码可读性
   - 移除多余空行

2. 依赖减少：
   - 移除strconv包依赖
   - 简化导入语句
   - 减少包复杂度

验证：
- ✅ 编译通过：go build ./...
- ✅ 测试通过：TestToolArgsValue所有测试用例
- ✅ 功能正确：类型安全访问功能正常
- ✅ 代码质量：代码更简洁，更易维护

预期效果：
- ✅ 类型安全：编译时类型检查，避免运行时错误
- ✅ 代码简洁：大幅简化复杂逻辑
- ✅ 性能优化：直接类型断言，减少开销
- ✅ 可维护性：清晰的代码结构，易于理解和修改

+++ internal/toolcall/code/code_search_semantic.go
@@ -241,9 +241,9 @@ func handleSearchCodeSemantic(ctx context.Context, args toolcall.ToolArgs) (outp
 	}
 
 	// 解析可选参数
-	contextLines := toolcall.ToolArgsValue(args, "context_lines", 5)
+	contextLines := int(toolcall.ToolArgsValue(args, "context_lines", int64(5)))
 	caseSensitive := toolcall.ToolArgsValue(args, "case_sensitive", false)
-	maxMatches := toolcall.ToolArgsValue(args, "max_matches", 0)
+	maxMatches := int(toolcall.ToolArgsValue(args, "max_matches", int64(0)))
 
 	// 扩展文件模式
 	files, err := file.ExpandFilePattern(filePattern)
