+++ internal/toolcall/issue/issue_close.go
@@ -9,7 +9,7 @@ import (
 
 // handleIssueClose 处理关闭issue（Tool Calling）
 func handleIssueClose(ctx context.Context, args toolcall.ToolArgs) (output string, user string, err error) {
-	number := toolcall.ToolArgsValue(args, "number", 0)
+	number := int(toolcall.ToolArgsValue(args, "number", int64(0)))
 	if number == 0 {
 		err = fmt.Errorf("必须提供issue编号")
 		return
