+++ internal/toolcall/issue/issue_reopen.go
@@ -30,7 +30,7 @@ func init() {
 
 // handleIssueReopen 处理重新打开issue（Tool Calling）
 func handleIssueReopen(ctx context.Context, args toolcall.ToolArgs) (output string, user string, err error) {
-	number := toolcall.ToolArgsValue(args, "number", 0)
+	number := int(toolcall.ToolArgsValue(args, "number", int64(0)))
 	if number == 0 {
 		err = fmt.Errorf("必须提供issue编号")
 		return
